requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn>=0.35.0,<1.0.0
requests>=2.32.0,<3.0.0
pillow>=11.3.0,<12.0.0
python-multipart>=0.0.20,<1.0.0
